backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-multipart
pdfplumber
python-docx
celery
redis
pydantic
email-validator
bcrypt
python-jose[cryptography]
python-dotenv
cryptography
pytesseract   # decommenta per OCR
pdf2image      # decommenta per OCR
Pillow         # decommenta per OCR